+++ src/module/SadnessChan.ts
@@ -78,16 +78,26 @@ class SadnessChan {
         const rolls = userData.rolls;
         if (rolls) {
             const crtFail = rolls[failNumber];
+            const averageNumber = this._getAverage(rolls);
             const crtSuccess = rolls[successNumber];
             const rollsClass = `${statsBodyClass}__rolls`;
             const rollClass = `${rollsClass}-roll`;
+            const averageRollDie = () => {
+                if (!Settings.getSetting(this._settingKeys.AVERAGE_TOGGLE)) return '';
+
+                return `<li class="${rollClass}">
+                    <span class="${rollClass}-dice avg"><span>${averageNumber}</span></span>    
+                    <span class="${rollClass}-count">avg</span>    
+                </li>`
+            };
 
             message += `
                 <ol class="${rollsClass}">
                     <li class="${rollClass}">
                         <span class="${rollClass}-dice min">${failNumber}</span>    
                         <span class="${rollClass}-count">${crtFail}</span>    
-                    </li>
+                    </li>    
+                    ${averageRollDie()}
                     <li class="${rollClass}">
                         <span class="${rollClass}-dice max">${successNumber}</span>    
                         <span class="${rollClass}-count">${crtSuccess}</span>
@@ -212,6 +222,21 @@ class SadnessChan {
     }
 
     /**
+     * Returns the average of all the rolls made by a user
+     *
+     * @private
+     */
+    private _getAverage(rolls: Array<number>): number {
+        let rollsTotal = 0;
+        let numberOfRolls = 0;
+        for (let i = 0; i < rolls.length; i++) {
+            rollsTotal += i * rolls[i];
+            numberOfRolls += rolls[i];
+        }
+        return Utils.roundUp(rollsTotal / numberOfRolls);
+    }
+
+    /**
      * Returns number of faces of the die based on user settings
      * The number of faces must be between 2 and 1000
      */
@@ -240,20 +265,27 @@ class SadnessChan {
      * Creates the stats message
      *
      * @param userData - current user
+     * @param displayPortrait -
      */
-    public getStatsMessage(userData: any): string {
+    public getStatsMessage(userData: any, displayPortrait: boolean = true): string {
         const statsClass = `${Utils.moduleName}-chat-stats`;
         const statsHeaderClass = `${statsClass}-header`;
-        const statsBodyClass = `${statsClass}-body`
+        const statsBodyClass = `${statsClass}-body`;
+
+        const portraitHTML = ()=> {
+            if (!displayPortrait) return '';
+
+            return `<div class="${statsHeaderClass}">
+                        ${this._getRandomPortrait(statsHeaderClass)}
+                        <h3 class="${statsHeaderClass}__name">
+                            ${Utils.moduleTitle}
+                        </h3>
+                    </div>`;
+        }
 
         return `
             <div class="${statsClass}">
-                <div class="${statsHeaderClass}">
-                    ${this._getRandomPortrait(statsHeaderClass)}
-                    <h3 class="${statsHeaderClass}__name">
-                        ${Utils.moduleTitle}
-                    </h3>
-                </div>
+                ${portraitHTML()}
                 <div class="${statsBodyClass}">
                     ${this._getStatsMessageBody(userData, statsBodyClass)}
                 </div>