+++ src/module/hooks/PreCreateChatMessage.ts
@@ -61,29 +61,53 @@ class PreCreateChatMessage {
         if (counter && counter[user]) {
             this._sendStatsMessage(message, options, counter[user], user);
             Utils.debug('Sad stats displayed.');
+        } else {
+            message.content = SadnessChan.generateMessageStructure(this._errorMessages.NO_DATA);
+            this._prepareMessage(message, options, user, true);
         }
     }
 
     private _sendAllRollsMessage(message: any, options: any, userId: any) {
         const counter = Settings.getCounter();
-        const rolls = counter[userId]?.rolls;
-        if (!(counter && rolls)) return;
+        if (!(counter && game.user.hasRole(4))) return;
 
-        message.content = rolls.reduce((result: string, el: number, index: number): string => {
-            return !index ? '' : result + `${index} ${el}<br>`;
-        }, '');
+        message.content = '';
+        const activeUsers = game.users.entities.filter((user) => user.active);
+        activeUsers.forEach((user, index) => {
+            // @ts-ignore
+            const userData = counter[user?.data?._id];
+            if (!userData) return;
+
+            message.content += SadnessChan.getStatsMessage(userData, index === 0);
+        })
 
         this._prepareMessage(message, options, userId);
     }
 
+    private _sendHelpMessage (message: any, options: any, userId: any) {
+        const command = SadnessChan.getCmd()
+        message.content = `<p>Are you that useless that you need help? Fine, I'll help you:</p>`;
+        message.content += `<p><b>${command}</b> - "happiness".</p>`
+        message.content += `<p><b>${command} all</b> - AOE "happiness".</p>`
+        message.content += `<p><b>${command} reset settings</b> - you want to make me a normie, sure I guess...</p>`
+        message.content += `<p><b>${command} reset lists</b> -  back to square one, like the retards who made me intended.</p>`
+        message.content += `<p><b>${command} reset counter</b> - makes me forget how much of a disappointment you are. Oh. Wait. I still have your browser history to remind me.</p>`
+        this._prepareMessage(message, options, userId);
+    }
+
+
     public executeCommand(args: string, user: any, message: any, options: any) {
         const resetCommand = 'reset';
         const allCommand = 'all';
+        const helpCommand = 'help';
         if (args.startsWith(resetCommand)) {
             return this._executeResetCmd(args.replace(resetCommand + ' ', ''), message, options, user);
         }
         if (args.startsWith(allCommand)) {
             return this._sendAllRollsMessage(message, options, user);
+        }
+        if (args.startsWith(helpCommand)) {
+            return this._sendHelpMessage(message, options, user);
         }
     }